Refactor graphql schema on backend
